backend/lambda/data-extractor/requirements.txt
boto3>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
